Foi atualizado os componentes e as páginas
