+++ src/components/formSideBar/index.js
@@ -3,6 +3,7 @@ import "./styles.scss";
 import "../img/repo.scss"
 import { useHistory } from 'react-router';
 import axios from 'axios';
+import Loading from '../../../src/components/loading';
 
 function SideBar(props) {
 
@@ -34,25 +35,27 @@ function SideBar(props) {
         history.push("/services");
     }
 
+    var [loading, setLoading] = useState(Loading);
 
     useEffect(() => {
+        setLoading(true)
         axios
             .get(`${reprografia_url}/myUser/`, {
                 headers: {
                     accessToken: localStorage.getItem("accessToken"),
                 },
             }).then((result) => {
-                if(result.data.roles) {
+                if (result.data.roles) {
                     setName(result.data.nome)
                     setNif(result.data.nif)
                     setImage(`${reprografia_url}/${result.data.imagem}`)
                     if (result.data.roles[0].descricao === "admin") {
                         setAdmin(true)
                     }
-                    else{
+                    else {
                         setAdmin(false)
                     }
-    
+
                     if (props.nif) {
                         setNif(props.nif)
                     }
@@ -62,39 +65,45 @@ function SideBar(props) {
                     if (props.name) {
                         setName(props.name)
                     }
-                    if(props.admin){
+                    if (props.admin) {
                         setAdmin(props.admin)
                     }
                 }
+                setLoading(false)
             })
-    }, [props.nif, props.image, props.name, props.admin])
+    }, [props.nif, props.image, props.name, props.admin, reprografia_url])
 
     return (
-        <div className="sidebarG">
-            <div onClick={() => { history.push(`/user/${nif}`) }} className="circle">
-                <img src={image} className="repo" alt="imagem do usuário" />
-            </div>
-            <h2 className="subTitle" onClick={() => { history.push(`/user/${nif}`) }}>{name}</h2>
-            <div className="buttonsG">
+        <>
+            {loading ? <Loading /> : <>
+                <div className="sidebarG">
+                    <div onClick={() => { history.push(`/user/${nif}`) }} className="circle">
+                        <img src={image} className="repo" alt="imagem do usuário" />
+                    </div>
+                    <h2 className="subTitle" onClick={() => { history.push(`/user/${nif}`) }}>{name}</h2>
+                    <h3 className="sidebar-nif">NIF: {nif}</h3>
+                    <div className="buttonsG">
 
-                {props.requestForm ? <></> : <><button className="buttonG" onClick={routeForm}>Solicitar Impressão</button></>}
-                {props.requestsNoInfo ? <></> :
-                    <>
-                        <button className="buttonG" onClick={routeMyRequests}>Meus Pedidos</button>
-                    </>
-                }
-                {admin ?
-                    <>
-                    
-                        {props.management ? <></> : <button className="buttonG" onClick={routeManagement}>Gerencia de usuários</button>}
-                        {props.estatisticas ? <></> : <button className="buttonG" onClick={routeStatistics}>Estatísticas</button>}
-                        {props.services ? <></> : <button className="buttonG" onClick={routeServices}>Serviços</button>}
-                    </> : 
-                    <>
-                    </>}
+                        {props.requestForm ? <></> : <><button className="buttonG" onClick={routeForm}>Solicitar Impressão</button></>}
+                        {props.requestsNoInfo ? <></> :
+                            <>
+                                <button className="buttonG" onClick={routeMyRequests}>Meus Pedidos</button>
+                            </>
+                        }
+                        {admin ?
+                            <>
 
-            </div>
-        </div>
+                                {props.management ? <></> : <button className="buttonG" onClick={routeManagement}>Gerencia de usuários</button>}
+                                {props.estatisticas ? <></> : <button className="buttonG" onClick={routeStatistics}>Estatísticas</button>}
+                                {props.services ? <></> : <button className="buttonG" onClick={routeServices}>Serviços</button>}
+                            </> :
+                            <>
+                            </>}
+
+                    </div>
+                </div>
+            </>}
+        </>
     );
 }
