+++ src/components/formSideBar/styles.scss
@@ -12,10 +12,17 @@
 
   .subTitle {
     cursor: pointer;
+    margin-top: 35rem;
+  }
+
+  .sidebar-nif {
+    color: $white;
+    font-size: 2rem;
+    margin-top: 1rem;
   }
 
   .buttonsG {
-    margin-top: 7rem;
+    margin-top: 2rem;
     display: flex;
     align-items: center;
     flex-direction: column;
