+++ src/components/img/repo.scss
@@ -4,7 +4,7 @@
   width: 16rem;
   height: 16rem;
   margin-top: 15rem;
-  position: relative;
+  position: absolute;
   background-color: $white;
   border-radius: 50%;
 }
