+++ src/pages/add-services/index.js
@@ -1,9 +1,8 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useHistory, useParams } from 'react-router-dom'
 import axios from 'axios';
 import '../../styles/addService.scss';
 import LoginContainer from '../../components/loginContainer';
-import Loading from '../../components/loading';
 
 export default function AddService() {
 
@@ -17,7 +16,7 @@ export default function AddService() {
   const [message, setMessage] = useState();
 
   const port = process.env.REACT_APP_PORT || 3002;
-  
+
   const reprografia_url = `${process.env.REACT_APP_REPROGRAFIA_URL}:${port}`;
 
   const AddService = () => {
@@ -44,87 +43,71 @@ export default function AddService() {
     })
   }
 
-  const voltar = () => {
-    axios
-      .get(`${reprografia_url}/auth`, {
-        headers: {
-          accessToken: localStorage.getItem("accessToken"),
-        },
-      }).then((result) => {
-        history.push("/services");
-      })
-  }
-
   const onSubmit = (e) => {
     e.preventDefault();
     AddService();
   }
 
-  var [loading, setLoading] = useState(Loading);
-
-  useEffect(() => {
-    setLoading(true);
-    setTimeout(() => {
-      setLoading(false);
-    }, 1300);
-  }, [])
 
   return (
     <>
-      {loading ? <> <Loading /> </> :
-        <>
-          <LoginContainer />
-          <div className="finishing">
-            <form onSubmit={onSubmit}>
-              <h2 id="h2" className="service-subTitle">
-                Adicionar Serviço
-              </h2>
-              <input
-                className="input-service"
-                name="descricao"
-                type="text"
-                placeholder="descrição de capa e acabamento"
-                required
-                onChange={(e) => {
-                  setDescricao(e.target.value);
-                }}
-              />
-              <input
-                className="input-service"
-                name="quantidade"
-                type="number"
-                placeholder="quantidade do serviço"
-                required
-                onChange={(e) => {
-                  setQuantidade(e.target.value);
-                }}
-              />
-              <input
-                className="input-service"
-                name="custo"
-                type="number"
-                step="any"
-                placeholder="custo unitário do serviço"
-                required
-                onChange={(e) => {
-                  setCusto(e.target.value);
-                }}
-              />
-              <h3>{message}</h3>
-              <input
-                type="submit"
-                className="nu-send-button"
-                id="btn"
-                value="Adicionar"
-              />
-            </form>
-            <button
-              className="btn-back-user"
-              id="btn"
-              onClick={voltar}>Voltar</button>
-          </div>
-        </>
-      }
+      <LoginContainer />
+      <div className="finishing">
+        <form onSubmit={onSubmit}>
+          <h2 id="" className="service-subTitle">
+            Adicionar Serviço
+          </h2>
+          <h4 className="import-addServices">Onde houver "*" o preenchimento é obrigatório</h4>
+          <label className="important">*
+            <input
+              className="input-service"
+              name="descricao"
+              type="text"
+              placeholder="descrição de capa e acabamento"
+              required
+              onChange={(e) => {
+                setDescricao(e.target.value);
+              }}
+            />
+          </label>
+          <label className="important">*
+            <input
+              className="input-service"
+              name="quantidade"
+              type="number"
+              placeholder="quantidade do serviço"
+              required
+              onChange={(e) => {
+                setQuantidade(e.target.value);
+              }}
+            />
+          </label>
+          <label className="important">*
+            <input
+              className="input-service"
+              name="custo"
+              type="number"
+              step="any"
+              placeholder="custo unitário do serviço"
+              required
+              onChange={(e) => {
+                setCusto(e.target.value);
+              }}
+            />
+          </label>
+          <h3>{message}</h3>
+          <input
+            type="submit"
+            className="nu-send-button"
+            id="btn"
+            value="Adicionar"
+          />
+        </form>
+        <button
+          className="btn-back-user"
+          id="btn"
+          onClick={() => { history.push("/services") }}>Voltar</button>
+      </div>
     </>
   );
 }
