+++ src/pages/feedbacks/index.js
@@ -9,125 +9,136 @@ import Menu from '../../../src/components/hamburgerButton';
 import SideBar from '../../../src/components/formSideBar';
 import Swal from 'sweetalert2'
 import Button from "@restart/ui/esm/Button";
+import Loading from "../../components/loading";
 
-function RequestList(props) {
+function Feedback(props) {
 
-  const { id } = useParams();
+    const { id } = useParams();
 
-  var [feedbacks, setFeedBacks] = useState({
-    status: false,
-    list: [],
-    message: "",
-  });
+    var [feedbacks, setFeedBacks] = useState({
+        status: false,
+        list: [],
+        message: "",
+    });
 
-  const Details = (feedbackId) => {
-    axios
-      .get(`http://localhost:3002/feedback/id/${feedbackId}`, {
-        headers: {
-          accessToken: localStorage.getItem("accessToken"),
-        },
-      })
-      .then((result) => {
-        console.log(result)
-        if (result.data.status === "error" || result.status === 500 || result.status === 403) {
-          Swal.fire(
-            `ERROR!`,
-            result.data.message
-          )
-        }
-        else {
-          Swal.fire(
-            `Detalhes da avaliação Nº${feedbackId}`,
-            `Solicitada pelo usuário com nif <strong>${result.data.userId}</strong> no pedido <strong>Nº${result.data.pedidoId}</strong> <br></br> <h3><strong>Avaliação:</strong></h3> ${result.data.avaliacao_obs}`,
-            'warning'
-          )
-        }
-      });
-  }
+    const port = process.env.REACT_APP_PORT || 3002;
 
-  useEffect(() => {
-    axios
-      .get(`http://localhost:3002/feedbacks/${id}`, {
-        headers: {
-          accessToken: localStorage.getItem("accessToken"),
-        },
-      })
-      .then((result) => {
-        console.log(result)
-        if (result.data.status === "error") {
-          setFeedBacks({
-            message: result.data.message
-          });
-        } else {
-          setFeedBacks({
-            list: result.data,
-            status: true,
-          });
-        }
-      });
-  }, [id]);
+    const reprografia_url = `${process.env.REACT_APP_REPROGRAFIA_URL}:${port}`;
 
-  return (
-    <>
-      <Menu />
-      <Header nif={props.nif} />
-      <SideBar image={props.image} name={props.name} admin={props.admin} />
+    const Details = (feedbackId) => {
+        axios
+            .get(`${reprografia_url}/feedback/id/${feedbackId}`, {
+                headers: {
+                    accessToken: localStorage.getItem("accessToken"),
+                },
+            })
+            .then((result) => {
+                console.log(result)
+                if (result.data.status === "error" || result.status === 500 || result.status === 403) {
+                    Swal.fire(
+                        `ERROR!`,
+                        result.data.message
+                    )
+                }
+                else {
+                    Swal.fire(
+                        `Detalhes da avaliação Nº${feedbackId}`,
+                        `Solicitada pelo usuário com nif <strong>${result.data.userId}</strong> no pedido <strong>Nº${result.data.pedidoId}</strong> <br></br> <h3><strong>Avaliação:</strong></h3> ${result.data.avaliacao_obs}`,
+                        'warning'
+                    )
+                }
+            });
+    }
+    var [loading, setLoading] = useState(Loading);
 
-      <div className="container-feedbacks">
-        <div className="titleFDB">
-          <h1>Avaliações do pedido Nº{id}</h1>
-        </div>
-      </div>
+    useEffect(() => {
+        setLoading(true)
+        axios
+            .get(`${reprografia_url}/feedbacks/${id}`, {
+                headers: {
+                    accessToken: localStorage.getItem("accessToken"),
+                },
+            })
+            .then((result) => {
+                if (result.data.status === "error") {
+                    setFeedBacks({
+                        message: result.data.message
+                    });
+                } else {
+                    setFeedBacks({
+                        list: result.data,
+                        status: true,
+                    });
+                }
+                setLoading(false)
+            });
+    }, [id, reprografia_url]);
 
-      <div className="container-feedbacks">
+    return (
         <>
-          {feedbacks.status ? (
-            <>
-              <Table className="table-feedbacks" striped bordered hover size="sm">
-                <thead>
-                  <th> Realizado por (nif): </th>
-                  <th>
-                    Realizado em:
-                  </th>
-                  <th>Status</th>
-                  <th>
-                    Detalhes
-                  </th>
-                </thead>
-                {feedbacks.list.map((data) => (
-                  <>
-                    <tbody>
-                      <tr>
-                        <td>
-                          <h1>{data.userId}</h1>
-                        </td>
+            {loading ? <Loading /> :
+                <>
+                    <Menu />
+                    <Header nif={props.nif} />
+                    <SideBar image={props.image} name={props.name} admin={props.admin} />
 
-                        <td>
-                          <h1>{data.createdAt}</h1>
-                        </td>
+                    <div className="container-feedbacks">
+                        <div className="titleFDB">
+                            <h1>Avaliações do pedido Nº{id}</h1>
+                        </div>
+                    </div>
 
-                        <td>
-                          {data.avaliacaoId === 1 ? <h1>Atendeu!</h1> : <h1>Não atendeu!</h1>}
-                        </td>
+                    <div className="container-feedbacks">
+                        <>
+                            {feedbacks.status ? (
+                                <>
+                                    <Table className="table-feedbacks" striped bordered hover size="sm">
+                                        <thead>
+                                            <th> Realizado por (nif): </th>
+                                            <th>
+                                                Realizado em:
+                                            </th>
+                                            <th>Status</th>
+                                            <th>
+                                                Detalhes
+                                            </th>
+                                        </thead>
+                                        {feedbacks.list.map((data) => (
+                                            <>
+                                                <tbody>
+                                                    <tr>
+                                                        <td>
+                                                            <h1>{data.userId}</h1>
+                                                        </td>
 
-                        <td>
-                          <Button className="btn-details" onClick={() => { Details(data.feedbackId) }}>detalhes</Button>
-                        </td>
-                      </tr>
-                    </tbody>
-                  </>
-                ))}
-              </Table>
-            </>
-          ) : (
-            <>
-              <h1>{feedbacks.message}</h1>
-            </>
-          )}
+                                                        <td>
+                                                            <h1>{data.createdAt}</h1>
+                                                        </td>
+
+                                                        <td>
+                                                            {data.avaliacaoId === 1 ? <h1>Atendeu!</h1> : <h1>Não atendeu!</h1>}
+                                                        </td>
+
+                                                        <td>
+                                                            <Button className="btn-details" onClick={() => { Details(data.feedbackId) }}>detalhes</Button>
+                                                        </td>
+                                                    </tr>
+                                                </tbody>
+                                            </>
+                                        ))}
+                                    </Table>
+                                </>
+                            ) : (
+                                <>
+                                    <h1>{feedbacks.message}</h1>
+                                </>
+                            )}
+                        </>
+                    </div>
+                </>
+            }
         </>
-      </div>
-    </>
-  );
+    );
 };
 
-export default RequestList;
+export default  Feedback;
