+++ src/pages/forgotPassword/index.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useHistory } from "react-router";
 import LoginContainer from '../../components/loginContainer';
 import '../../styles/forgotPassword.scss';
