+++ src/styles/newUser.scss
@@ -29,15 +29,33 @@ form {
   }
 }
 
-.input-box {
-  border: 1px solid #363232;
-  box-sizing: border-box;
-  border-radius: 4px;
-  margin-bottom: 1.8rem;
-  height: 3.5rem;
+.selectNew {
+  display: flex;
+  justify-content: center;
+  font-size: 1.5rem;
+  border: solid 2px $red;
+  padding: 0.5rem;
   width: 38rem;
-  padding-left: 0.5rem;
+  margin: auto;
 }
+
+.important {
+  color: $red;
+  font-size: 2.3rem;
+  position: static;
+  margin-right: 0.5rem;
+
+  .input-box {
+    border: 2px solid $red;
+    box-sizing: border-box;
+    border-radius: 4px;
+    margin-bottom: 1.8rem;
+    height: 3.5rem;
+    width: 38rem;
+    padding-left: 0.5rem;
+  }
+}
+
 
 .nu-subTitle {
   display: flex;
@@ -60,7 +78,7 @@ form {
 }
 
 .nu-send-button {
-  width: 38rem;
+  width: 40rem;
   height: 3.5rem;
   cursor: pointer;
   margin-top: 2rem;
@@ -79,8 +97,8 @@ form {
   }
 }
 
-.nu-back-button {
-  width: 38rem;
+#btn.btn-back-user {
+  width: 40rem;
   height: 3.5rem;
   cursor: pointer;
   margin-top: 2rem;
@@ -98,6 +116,32 @@ form {
   }
 }
 
+.customizeNew:hover {
+  background-color: #fb0404;
+}
+.customizeNew {
+  background-color: #4895ef;
+  width: 100%;
+  color: #ffffff;
+  font-weight: bold;
+  font-size: 1.4rem;
+  border-radius: 16px;
+  border: 0;
+  padding: 0.6rem 12.5rem;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  transition: 0.5s;
+  text-align: center;
+  margin-bottom: 2rem;
+  margin-top: 2rem;
+}
+
+.customizeNew .uploudNew {
+  margin-right: 2rem;
+  font-size: 2rem;
+}
+
 .input-title {
   display: flex;
   align-items: center;
@@ -105,7 +149,7 @@ form {
   background-color: $grey-button;
   color: $white;
   height: 2.5rem;
-  width: 38rem;
-  margin-bottom: 1.5rem;
+  width: 40rem;
+  margin-bottom: 2rem;
   font-size: 1.8rem;
 }
