+++ src/styles/requestForm.scss
@@ -1,4 +1,4 @@
-@import "../../styles/global";
+@import "_global.scss";
 
 .containerForm {
   display: flex;
